move maze generator to maze file
